+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Unreleased
+
+### @routo/core
+
+- Added support for trailing slash in `basename` option
+
 ## 0.6.0 (June 2, 2020)
 
 ### @routo/core